+++ src/ringbuffer/ringbuffer.h
@@ -1,6 +1,8 @@
 #ifndef _RINGBUFFER_H_
 #define _RINGBUFFER_H_
 
+#include <cstddef>
+#include <cstdint>
 #include <cstring>
 
 int RingBufferDummy();
@@ -9,7 +11,7 @@ class Data {
  public:
   Data() = default;
 
-  Data(const char* data, size_t size) {
+  Data(const char* data, uint32_t size) {
     data_ = new char[size];
     memcpy(data_, data, size);
     size_ = size;
@@ -36,12 +38,12 @@ class Data {
     size_ = 0;
   }
 
-  size_t size() const { return size_; }
+  uint32_t size() const { return size_; }
   char* data() const { return data_; }
 
  public:
   char* data_ = nullptr;
-  size_t size_ = 0;
+  uint32_t size_ = 0;
 };
 
 template <typename T>
@@ -101,8 +103,8 @@ class RingBuffer {
 
  private:
   unsigned int m_size;
-  int m_front;
-  int m_rear;
+  unsigned int m_front;
+  unsigned int m_rear;
   T* m_data;
 };
